Only run the unit tests which touch changed code in the ci (#96)
Redo of https://github.com/struphy-hub/struphy/pull/50

**Solves the following issue(s):**

Closes https://github.com/struphy-hub/struphy/issues/45

**Core changes:**

There are now four PR workflows that must pass for a feature to be
merged:

1. unit
2. models
3. pure Python
4. with clones

Tests are made with and without MPI with `--testmon`. Testmon data is
cached for the next workflow run to use. Cache cannot be altered, thus
new cache keys are necessary in every run. The key convention is, for
unit tests for example:
```
key: testmon-unit-${{ github.event.number }}-${{ github.run_number }}
restore-keys: |
    testmon-unit-${{ github.event.number }}-
    testmon-unit-
``` 
The `event.number` is the number of the PR. The `run_number` refers to
the workflow run, thus it increases monotonically. The one that is
fetched has a run number one less than the current run, and is thus the
most recent.

---------

Co-authored-by: Stefan Possanner <stefan.possanner@ipp.mpg.de>
Co-authored-by: Stefan Possanner <86720346+spossann@users.noreply.github.com>